requirements.txt
absl-py==1.3.0
aspose-words==22.12.0
astunparse==1.6.3
boto==2.49.0
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.5
cycler==0.11.0
Flask==2.2.2
Flask-Cors==3.0.10
flatbuffers==22.10.26
fonttools==4.37.4
future==0.18.2
gast==0.4.0
google-auth==2.14.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.50.0
h5py==3.7.0
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
llvmlite==0.39.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.1
numba==0.56.3
numpy==1.23.3
oauthlib==3.2.2
opencv-python==4.6.0.66
openpifpaf==0.13.5
opt-einsum==3.3.0
packaging==21.3
Pillow==9.2.0
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
pysparkling==0.6.1
python-dateutil==2.8.2
python-json-logger==2.0.4
pytz==2022.4
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.1
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.1.0
torch==1.12.1
torchvision==0.13.1
typing-extensions==4.4.0
urllib3==1.26.12
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.9.0
